requirements.txt
flask
pyspark
pymysql
psycopg2-binary
mysql-connector-python
pymongo
